+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Calling `eyes.check` with `fully(false)` now doesn't take full page screenshot even if `forceFullPageScreenshot` is set to true in the configuration. [Trello 1926](https://trello.com/c/4vcerUTm)
 - `saveDebugScreenshot` now works when there is no full page screenshot. [Trello 1138](https://trello.com/c/DUjJxuMH)
 - Removed unused `Target` class in the java.sdk.core. [Trello 1098](https://trello.com/c/Oi36yIro)
+- Added missing constructors to `com.applitools.eyes.selenium.Configuration`. [Trello 1954](https://trello.com/c/al9nZGPD)
 ### Updated
 - Updated the API of the IOS Simulation. [Trello 1944](https://trello.com/c/EzyG7525)